+++ frontend/src/features/DevAuth/DevAuth.tsx
@@ -1,158 +1,14 @@
-import {
-  FTH1,
-  FTH3,
-  FTH4,
-  FTButton,
-  FTSubmit,
-  FTTextField,
-} from '@/components/FTBasicComponents';
-import { useQuery, useStoredCredential } from '@/hooks';
-import { useState, useEffect, useMemo } from 'react';
+import { AppCredential, useQuery } from '@/hooks';
+import { useState, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
+import {
+  DevAuthenticated,
+  DevAuthLogin,
+  DevAuthValidating,
+  UserPersonalData,
+} from './AuthCard';
 
 const apiHost = `http://localhost:3000`;
-
-const FtAuthForm = () => (
-  <form method="POST" action="http://localhost:3000/auth/login_ft">
-    <FTSubmit className="hover:bg-white hover:text-black" value="Login" />
-  </form>
-);
-
-const SelfAuthForm = (props: { finalizer: (data: any) => void }) => {
-  const [userIdStr, setUserIdStr] = useState('');
-  type Phase = 'Ready' | 'NotReady' | 'Working';
-  // 状態遷移
-  // - ボタン押せる
-  // - ボタン押せない
-  // - ボタン押してる
-  const [phase, setPhase] = useState<Phase>('NotReady');
-
-  const sender = async (s: string) => {
-    const url = `${apiHost}/auth/self/${s}`;
-    const result = await fetch(url, {
-      method: 'GET',
-      mode: 'cors',
-    });
-    const json = await result.json();
-    console.log('json', json);
-    props.finalizer(json);
-  };
-
-  const validator = (s: string) => {
-    if (!s) {
-      return 'empty?';
-    }
-    const n = parseInt(s);
-    if (`${n}` !== s) {
-      return 'not a valid number?';
-    }
-    if (n < 1) {
-      return 'not a valid userId?';
-    }
-    return null;
-  };
-
-  const click = async () => {
-    try {
-      setPhase('Working');
-      await sender(userIdStr);
-    } catch (e) {
-      console.error(e);
-    }
-    setPhase('Ready');
-  };
-
-  const errorMessage = (() => {
-    if (phase === 'Working') {
-      return null;
-    }
-    return validator(userIdStr);
-  })();
-
-  return (
-    <>
-      <FTTextField
-        className="border-2"
-        autoComplete="off"
-        placeholder="ユーザID"
-        value={userIdStr}
-        onChange={(e) => setUserIdStr(e.target.value)}
-      />
-      <FTButton disabled={!!errorMessage} onClick={() => click()}>
-        Force Login
-      </FTButton>
-      <div>{errorMessage}</div>
-    </>
-  );
-};
-
-export const DevAuthLogin = (props: { finalizer: (data: any) => void }) => {
-  return (
-    <>
-      <FTH1 className="text-4xl font-bold" style={{ padding: '4px' }}>
-        Login, Please.
-      </FTH1>
-      <br />
-      <div className="flex flex-col gap-8">
-        <FTH3>By 42Auth</FTH3>
-        <div className="text-center">
-          <FtAuthForm />
-        </div>
-        <FTH3>By Self</FTH3>
-        <div className="text-center">
-          <SelfAuthForm finalizer={props.finalizer} />
-        </div>
-        <FTH3>By Email / Password</FTH3>
-        <div></div>
-      </div>
-    </>
-  );
-};
-
-export const DevAuthenticated = (props: {
-  personalData: UserPersonalData;
-  onLogout?: () => void;
-}) => {
-  return (
-    <>
-      <FTH1 className="text-4xl font-bold" style={{ padding: '4px' }}>
-        You&apos;re Authenticated.
-      </FTH1>
-      <div className="flex flex-col gap-2">
-        <FTH4>id</FTH4>
-        <div>{props.personalData.id}</div>
-        <FTH4>name</FTH4>
-        <div>{props.personalData.displayName}</div>
-        <FTH4>email</FTH4>
-        <div>{props.personalData.email}</div>
-      </div>
-      <br />
-      <div className="flex flex-col gap-2">
-        <FTH4>Logout?</FTH4>
-        <div className="text-center">
-          <FTButton
-            onClick={() => {
-              if (props.onLogout) props.onLogout();
-            }}
-          >
-            Logout
-          </FTButton>
-        </div>
-      </div>
-    </>
-  );
-};
-
-export const DevAuthValidating = () => {
-  return (
-    <>
-      <FTH1 className="text-4xl font-bold" style={{ padding: '4px' }}>
-        Just A Moment, Please.
-      </FTH1>
-      <div>クレデンシャルを検証しています...</div>
-    </>
-  );
-};
 
 type AuthenticationFlowState =
   | 'Neutral'
@@ -161,22 +17,18 @@ type AuthenticationFlowState =
   | 'NotAuthenticated'
   | 'NeutralAuthorizationCode'
   | 'ValidatingAuthorizationCode';
-type UserPersonalData = {
-  id: number;
-  email: string;
-  displayName: string;
-};
 
-let i = 0;
-
-export const DevAuth = () => {
+export const DevAuth = (props: {
+  storedCredential: AppCredential | null;
+  setStoredCredential: (val: AppCredential | null) => void;
+  personalData: UserPersonalData | null;
+  setPersonalData: (val: UserPersonalData | null) => void;
+}) => {
   const query = useQuery();
   const navigation = useNavigate();
-  const j = i++;
   // 認証フローの状態
   const [initialAuthCode, initialFlowState] = (() => {
     const code = query.get('code');
-    console.log(j, { code });
     if (!code || typeof code !== 'string') {
       return ['', 'Neutral'] as const;
     }
@@ -185,43 +37,66 @@ export const DevAuth = () => {
   })();
   const [authState, setAuthState] =
     useState<AuthenticationFlowState>(initialFlowState);
-  // ブラウザが保持しているクレデンシャル
-  const [storedCredential, setStoredCredential] = useStoredCredential();
   // 認可コード
   const [ftAuthCode] = useState(initialAuthCode);
-  // パーソナルデータ
-  const [personalData, setPersonalData] = useState<UserPersonalData | null>(
-    null
-  );
 
-  const callSession = async () => {
-    console.log({ storedCredential });
-    if (storedCredential && storedCredential.token) {
+  const callSession = async (
+    onSucceeded: (user: any) => void,
+    onFailed: () => void
+  ) => {
+    console.log({ storedCredential: props.storedCredential });
+    if (props.storedCredential && props.storedCredential.token) {
       console.log(`calling callSession`);
       try {
         const result = await fetch(`${apiHost}/auth/session`, {
           method: 'GET',
           mode: 'cors',
           headers: {
-            Authorization: `Bearer ${storedCredential.token}`,
+            Authorization: `Bearer ${props.storedCredential.token}`,
           },
         });
         const json = await result.json();
         console.log('callSession', json);
         if (json) {
-          setPersonalData(json);
-          setAuthState('Authenticated');
+          onSucceeded(json);
           return;
         }
       } catch (e) {
         console.error(e);
       }
     }
-    setPersonalData(null);
+    onFailed();
+  };
+
+  const doLogout = () => {
+    props.setStoredCredential(null);
+    props.setPersonalData(null);
     setAuthState('NotAuthenticated');
   };
 
-  const callCallbackFt = async () => {
+  const finalizer = (token: string, user: any) => {
+    props.setStoredCredential({ token });
+    props.setPersonalData(user);
+    setAuthState('Authenticated');
+  };
+
+  const invokeSession = () =>
+    callSession(
+      (user) => {
+        props.setPersonalData(user);
+        setAuthState('Authenticated');
+      },
+      () => {
+        props.setPersonalData(null);
+        setAuthState('NotAuthenticated');
+      }
+    );
+  const invokeCallbackFt = () => callCallbackFt(finalizer, doLogout);
+
+  const callCallbackFt = async (
+    onSucceeded: (token: string, user: any) => void,
+    onFailed: () => void
+  ) => {
     const url = `${apiHost}/auth/callback_ft?code=${ftAuthCode}`;
     console.log(`calling callCallbackFt`, url);
     const result = await fetch(url, {
@@ -229,24 +104,22 @@ export const DevAuth = () => {
       mode: 'cors',
       headers: {},
     });
-    const { access_token, user } = (await result.json()) || {};
-    if (access_token && typeof access_token === 'string') {
-      // アクセストークンが得られた
-      // -> 認証完了状態にする
-      setStoredCredential({ token: access_token });
-      setPersonalData(user);
-      setAuthState('Authenticated');
-    } else {
-      // アクセストークンがなかった
-      // クレデンシャルを破棄する
-      setStoredCredential(null);
-      setAuthState('NotAuthenticated');
+    if (result.ok) {
+      const { access_token, user } = (await result.json()) || {};
+      if (access_token && typeof access_token === 'string') {
+        // アクセストークンが得られた
+        // -> 認証完了状態にする
+        onSucceeded(access_token, user);
+        return;
+      }
     }
+    // アクセストークンがなかった
+    // クレデンシャルを破棄する
+    onFailed();
   };
 
   // 認証状態のチェック
   useEffect(() => {
-    console.log(`authState[${j}]`, authState);
     switch (authState) {
       case 'Neutral': {
         // [検証中]
@@ -255,7 +128,7 @@ export const DevAuth = () => {
         break;
       }
       case 'Validating':
-        callSession();
+        invokeSession();
         break;
       case 'NotAuthenticated':
         // [未認証]
@@ -272,23 +145,11 @@ export const DevAuth = () => {
         break;
       }
       case 'ValidatingAuthorizationCode': {
-        callCallbackFt();
+        invokeCallbackFt();
         break;
       }
     }
   }, [authState]);
-
-  const doLogout = () => {
-    setStoredCredential(null);
-    setAuthState('NotAuthenticated');
-  };
-
-  const finalizer = (data: any) => {
-    const { token, user } = data;
-    setStoredCredential({ token });
-    setPersonalData(user);
-    setAuthState('Authenticated');
-  };
 
   const presentator = (() => {
     switch (authState) {
@@ -297,7 +158,7 @@ export const DevAuth = () => {
         return DevAuthValidating();
       case 'Authenticated':
         return DevAuthenticated({
-          personalData: personalData!,
+          personalData: props.personalData!,
           onLogout: doLogout,
         });
       case 'NotAuthenticated':
